Use correct cors config with security
