--- src/main/kotlin/three/consulting/epoc/config/CorsConfig.kt
@@ -1,28 +1,0 @@
-package three.consulting.epoc.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.http.HttpMethod
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class CorsConfig {
-
-    @Bean
-    fun webMvcConfigurer(): WebMvcConfigurer =
-        object : WebMvcConfigurer {
-            override fun addCorsMappings(registry: CorsRegistry) {
-                registry.addMapping("/**")
-                    .allowedOrigins("*")
-                    .allowedMethods(
-                        HttpMethod.GET.name,
-                        HttpMethod.POST.name,
-                        HttpMethod.PUT.name,
-                        HttpMethod.DELETE.name
-                    )
-                    .allowCredentials(false)
-                    .maxAge(3600)
-            }
-        }
-}
