+++ src/main/kotlin/three/consulting/epoc/config/SecurityConfig.kt
@@ -1,16 +1,21 @@
 package three.consulting.epoc.config
 
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
+import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
 @Profile("default")
+@Configuration
 @EnableWebSecurity
 class SecurityConfig : WebSecurityConfigurerAdapter() {
 
     override fun configure(http: HttpSecurity) {
-        http.authorizeRequests {
+        http.cors().and().authorizeRequests {
             it.mvcMatchers("/docs", "/docs/**", "/docs-ui.html", "/swagger-ui/**")
                 .permitAll()
                 .anyRequest()
@@ -19,4 +24,16 @@ class SecurityConfig : WebSecurityConfigurerAdapter() {
                 .oauth2ResourceServer().jwt()
         }
     }
+
+    @Bean
+    fun webMvcConfigurer(): WebMvcConfigurer =
+        object : WebMvcConfigurer {
+            override fun addCorsMappings(registry: CorsRegistry) {
+                registry.addMapping("/**")
+                    .allowedOriginPatterns("*")
+                    .allowedMethods("*")
+                    .allowCredentials(true)
+                    .maxAge(3600)
+            }
+        }
 }
